refactor: move utils into libs
